Mudando estrutura de arquivos dos components
